services/pricing/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
eth-abi==4.2.1
eth-utils==2.3.1
eth-keys==0.5.1
pytest==8.3.2
